refactor(routing): simplify historyRouter
