+++ packages/instantsearch-core/src/routing/__tests__/historyRouter-integration.test.ts
@@ -17,126 +17,128 @@ beforeEach(() => {
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-test('keeps url with cleanUrlOnDispose: false', async () => {
-  const router = historyRouter({ writeDelay, cleanUrlOnDispose: false });
+describe('historyRouter({ cleanUrlOnDispose })', () => {
+  test('keeps url with cleanUrlOnDispose: false', async () => {
+    const router = historyRouter({ writeDelay, cleanUrlOnDispose: false });
 
-  const indexName = 'indexName';
-  const search = instantsearch({
-    indexName,
-    searchClient: createSearchClient(),
-    routing: {
-      router,
-    },
+    const indexName = 'indexName';
+    const search = instantsearch({
+      indexName,
+      searchClient: createSearchClient(),
+      routing: {
+        router,
+      },
+    });
+
+    search.addWidgets([
+      connectPagination(() => {})({}),
+      index({ indexName }).addWidgets([connectSearchBox(() => {})({})]),
+    ]);
+
+    search.start();
+
+    expect(window.location.search).toBe('');
+
+    // on nested index
+    search.renderState[indexName].searchBox!.refine('test');
+    // on main index
+    search.renderState[indexName].pagination!.refine(39);
+
+    await wait(writeWait);
+
+    expect(window.location.search).toBe(
+      `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
+    );
+
+    search.dispose();
+
+    await wait(writeWait);
+
+    // URL has not been cleaned
+    expect(window.location.search).toBe(
+      `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
+    );
   });
 
-  search.addWidgets([
-    connectPagination(() => {})({}),
-    index({ indexName }).addWidgets([connectSearchBox(() => {})({})]),
-  ]);
+  test('clears url with cleanUrlOnDispose: true', async () => {
+    const router = historyRouter({ writeDelay, cleanUrlOnDispose: true });
 
-  search.start();
+    const indexName = 'indexName';
+    const search = instantsearch({
+      indexName,
+      searchClient: createSearchClient(),
+      routing: {
+        router,
+      },
+    });
 
-  expect(window.location.search).toBe('');
+    search.addWidgets([
+      connectPagination(() => {})({}),
+      index({ indexName }).addWidgets([connectSearchBox(() => {})({})]),
+    ]);
 
-  // on nested index
-  search.renderState[indexName].searchBox!.refine('test');
-  // on main index
-  search.renderState[indexName].pagination!.refine(39);
+    search.start();
 
-  await wait(writeWait);
+    expect(window.location.search).toBe('');
 
-  expect(window.location.search).toBe(
-    `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
-  );
+    // on nested index
+    search.renderState[indexName].searchBox!.refine('test');
+    // on main index
+    search.renderState[indexName].pagination!.refine(39);
 
-  search.dispose();
+    await wait(writeWait);
 
-  await wait(writeWait);
+    expect(window.location.search).toBe(
+      `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
+    );
 
-  // URL has not been cleaned
-  expect(window.location.search).toBe(
-    `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
-  );
-});
+    search.dispose();
 
-test('clears url with cleanUrlOnDispose: true', async () => {
-  const router = historyRouter({ writeDelay, cleanUrlOnDispose: true });
+    await wait(writeWait);
 
-  const indexName = 'indexName';
-  const search = instantsearch({
-    indexName,
-    searchClient: createSearchClient(),
-    routing: {
-      router,
-    },
+    // URL has been cleaned
+    expect(window.location.search).toBe('');
   });
 
-  search.addWidgets([
-    connectPagination(() => {})({}),
-    index({ indexName }).addWidgets([connectSearchBox(() => {})({})]),
-  ]);
+  test('does not clear url with cleanUrlOnDispose: undefined', async () => {
+    const router = historyRouter({ writeDelay });
 
-  search.start();
+    const indexName = 'indexName';
+    const search = instantsearch({
+      indexName,
+      searchClient: createSearchClient(),
+      routing: {
+        router,
+      },
+    });
 
-  expect(window.location.search).toBe('');
+    search.addWidgets([
+      connectPagination(() => {})({}),
+      index({ indexName }).addWidgets([connectSearchBox(() => {})({})]),
+    ]);
 
-  // on nested index
-  search.renderState[indexName].searchBox!.refine('test');
-  // on main index
-  search.renderState[indexName].pagination!.refine(39);
+    search.start();
 
-  await wait(writeWait);
+    expect(window.location.search).toBe('');
 
-  expect(window.location.search).toBe(
-    `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
-  );
+    // on nested index
+    search.renderState[indexName].searchBox!.refine('test');
+    // on main index
+    search.renderState[indexName].pagination!.refine(39);
 
-  search.dispose();
+    await wait(writeWait);
 
-  await wait(writeWait);
+    expect(window.location.search).toBe(
+      `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
+    );
 
-  // URL has been cleaned
-  expect(window.location.search).toBe('');
-});
+    search.dispose();
 
-test('does not clear url with cleanUrlOnDispose: undefined', async () => {
-  const router = historyRouter({ writeDelay });
+    await wait(writeWait);
 
-  const indexName = 'indexName';
-  const search = instantsearch({
-    indexName,
-    searchClient: createSearchClient(),
-    routing: {
-      router,
-    },
+    // URL has not been cleaned
+    expect(window.location.search).toBe(
+      `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
+    );
   });
-
-  search.addWidgets([
-    connectPagination(() => {})({}),
-    index({ indexName }).addWidgets([connectSearchBox(() => {})({})]),
-  ]);
-
-  search.start();
-
-  expect(window.location.search).toBe('');
-
-  // on nested index
-  search.renderState[indexName].searchBox!.refine('test');
-  // on main index
-  search.renderState[indexName].pagination!.refine(39);
-
-  await wait(writeWait);
-
-  expect(window.location.search).toBe(
-    `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
-  );
-
-  search.dispose();
-
-  await wait(writeWait);
-
-  // URL has not been cleaned
-  expect(window.location.search).toBe(
-    `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
-  );
 });
