+++ packages/instantsearch-core/src/routing/__tests__/routing-clean/dispose-start.test.ts
@@ -5,20 +5,29 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import { instantsearch, historyRouter, connectSearchBox } from '../..';
+import { instantsearch, historyRouter, connectSearchBox } from '../../..';
+
+import type { InstantSearch } from '../../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-describe('routing with external influence', () => {
-  test('keeps on working when the URL is updated by another program', async () => {
+const addWidgetsAndStart = (search: InstantSearch) => {
+  search.addWidgets([connectSearchBox(() => {})({})]);
+  search.start();
+};
+
+describe('routing back and forth to an InstantSearch instance', () => {
+  test('updates the URL after the instance is disposed then restarted', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. External influence: '/about'
-    // 4. Refine: '/about?indexName[query]=Samsung'
+    // 3. Dispose: '/'
+    // 4. Refine: '/'
+    // 5. Start: '/'
+    // 6. Refine: '/?indexName[query]=Samsung'
 
     const pushState = jest.spyOn(window.history, 'pushState');
 
@@ -27,6 +36,7 @@ describe('routing with external influence', () => {
       searchClient: createSearchClient(),
       routing: {
         router: historyRouter({
+          cleanUrlOnDispose: true,
           writeDelay,
         }),
       },
@@ -34,8 +44,7 @@ describe('routing with external influence', () => {
 
     // 1. Initial: '/'
     {
-      search.addWidgets([connectSearchBox(() => {})({})]);
-      search.start();
+      addWidgetsAndStart(search);
 
       await wait(writeWait);
       expect(window.location.search).toEqual('');
@@ -53,25 +62,42 @@ describe('routing with external influence', () => {
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. External influence: '/about'
+    // 3. Dispose: '/'
     {
-      window.history.pushState({}, '', '/about');
+      search.dispose();
 
       await wait(writeWait);
-      expect(window.location.pathname).toEqual('/about');
       expect(window.location.search).toEqual('');
       expect(pushState).toHaveBeenCalledTimes(2);
     }
 
-    // 4. Refine: '/about?indexName[query]=Samsung'
+    // 4. Refine: '/'
     {
       search.renderState.indexName.searchBox!.refine('Samsung');
 
       await wait(writeWait);
-      expect(window.location.pathname).toEqual('/about');
+      expect(window.location.search).toEqual('');
+      expect(pushState).toHaveBeenCalledTimes(2);
+    }
+
+    // 5. Start: '/'
+    {
+      addWidgetsAndStart(search);
+
+      await wait(writeWait);
+      expect(window.location.search).toEqual('');
+      expect(pushState).toHaveBeenCalledTimes(2);
+    }
+
+    // 6. Refine: '/?indexName[query]=Samsung'
+    {
+      search.renderState.indexName.searchBox!.refine('Samsung');
+
+      await wait(writeWait);
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Samsung')}`
       );
+      expect(pushState).toHaveBeenCalledTimes(3);
     }
   });
 });
