+++ packages/instantsearch-core/src/routing/__tests__/routing-clean/external-influence.test.ts
@@ -5,19 +5,20 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import { instantsearch, historyRouter, connectSearchBox } from '../..';
+import { instantsearch, historyRouter, connectSearchBox } from '../../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-describe('routing with no navigation', () => {
-  test('does not clean the URL when InstantSearch is disposed within the same page', async () => {
+describe('routing with external influence', () => {
+  test('keeps on working when the URL is updated by another program', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. Dispose: '/?indexName[query]=Apple'
+    // 3. External influence: '/about'
+    // 4. Refine: '/about?indexName[query]=Samsung'
 
     const pushState = jest.spyOn(window.history, 'pushState');
 
@@ -26,6 +27,7 @@ describe('routing with no navigation', () => {
       searchClient: createSearchClient(),
       routing: {
         router: historyRouter({
+          cleanUrlOnDispose: true,
           writeDelay,
         }),
       },
@@ -46,23 +48,31 @@ describe('routing with no navigation', () => {
       search.renderState.indexName.searchBox!.refine('Apple');
 
       await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Apple')}`
       );
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. Dispose: '/'
+    // 3. External influence: '/about'
     {
-      search.dispose();
+      window.history.pushState({}, '', '/about');
 
       await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
+      expect(window.location.pathname).toEqual('/about');
+      expect(window.location.search).toEqual('');
+      expect(pushState).toHaveBeenCalledTimes(2);
+    }
+
+    // 4. Refine: '/about?indexName[query]=Samsung'
+    {
+      search.renderState.indexName.searchBox!.refine('Samsung');
+
+      await wait(writeWait);
+      expect(window.location.pathname).toEqual('/about');
       expect(window.location.search).toEqual(
-        `?${encodeURI('indexName[query]=Apple')}`
+        `?${encodeURI('indexName[query]=Samsung')}`
       );
-      expect(pushState).toHaveBeenCalledTimes(1);
     }
   });
 });
