+++ packages/instantsearch-core/src/routing/__tests__/routing-clean/spa-replace-state.test.ts
@@ -5,22 +5,24 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import { instantsearch, historyRouter, connectSearchBox } from '../..';
+import { instantsearch, historyRouter, connectSearchBox } from '../../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-describe('routing with third-party client-side router', () => {
-  test('does not clean the URL after navigating', async () => {
+describe('routing using `replaceState`', () => {
+  // We can't assert whether another router did update the URL
+  // So there's no way to prevent `write` after `dispose`
+  test('cleans the URL after navigating', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. Navigate: '/about'
-    // 4. Back: '/?indexName[query]=Apple'
-    // 5. Restart: '/?indexName[query]=Apple'
-    // 6. Refine: '/?indexName[query]=Samsung'
+    // 3. Dispose: does not yet write
+    // 4. Route change (with `replaceState`): '/about?external=true', replaces state 2
+    // 5. Dispose: writes '/about' (this is a bug, and should be fixed when we have a way to prevent it)
+    // 6. Back: '/about?external=true'
 
     const pushState = jest.spyOn(window.history, 'pushState');
 
@@ -29,8 +31,8 @@ describe('routing with third-party client-side router', () => {
       searchClient: createSearchClient(),
       routing: {
         router: historyRouter({
-          writeDelay,
           cleanUrlOnDispose: true,
+          writeDelay,
         }),
       },
     });
@@ -56,49 +58,31 @@ describe('routing with third-party client-side router', () => {
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. Navigate: '/about'
+    // 3. Dispose: '/about'
+    // 4. Route change (with `replaceState`): '/about'
     {
       search.dispose();
-      window.history.pushState({}, '', '/about');
+      window.history.replaceState({}, '', '/about?external=true');
 
+      // Asserting `replaceState` call
+      expect(window.location.pathname).toEqual('/about');
+      expect(window.location.search).toEqual('?external=true');
+      expect(pushState).toHaveBeenCalledTimes(1);
+
+      // Asserting `dispose` calling `pushState`
       await wait(writeWait);
       expect(window.location.pathname).toEqual('/about');
       expect(window.location.search).toEqual('');
       expect(pushState).toHaveBeenCalledTimes(2);
     }
 
-    // 4. Back to previous page
+    // 5. Back: '/about'
     {
       window.history.back();
 
       await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
-      expect(window.location.search).toEqual(
-        `?${encodeURI('indexName[query]=Apple')}`
-      );
-    }
-
-    // 5. Restart InstantSearch
-    {
-      search.addWidgets([connectSearchBox(() => {})({})]);
-      search.start();
-
-      await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
-      expect(window.location.search).toEqual(
-        `?${encodeURI('indexName[query]=Apple')}`
-      );
-    }
-
-    // 6. Refine: '/?indexName[query]=Samsung'
-    {
-      search.renderState.indexName.searchBox!.refine('Samsung');
-
-      await wait(writeWait);
-      expect(window.location.search).toEqual(
-        `?${encodeURI('indexName[query]=Samsung')}`
-      );
-      expect(pushState).toHaveBeenCalledTimes(3);
+      expect(window.location.pathname).toEqual('/about');
+      expect(window.location.search).toEqual('?external=true');
     }
   });
 });
