+++ packages/instantsearch-core/src/routing/__tests__/routing-clean/spa.test.ts
@@ -5,28 +5,21 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import { instantsearch, historyRouter, connectSearchBox } from '../..';
-
-import type { InstantSearch } from '../..';
+import { instantsearch, historyRouter, connectSearchBox } from '../../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-const addWidgetsAndStart = (search: InstantSearch) => {
-  search.addWidgets([connectSearchBox(() => {})({})]);
-  search.start();
-};
-
-describe('routing back and forth to an InstantSearch instance', () => {
-  test('updates the URL after the instance is disposed then restarted', async () => {
+describe('routing with third-party client-side router', () => {
+  test('does not clean the URL after navigating', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. Dispose: '/?indexName[query]=Apple'
-    // 4. Refine: '/?indexName[query]=Apple'
-    // 5. Start: '/?indexName[query]=Apple'
+    // 3. Navigate: '/about'
+    // 4. Back: '/?indexName[query]=Apple'
+    // 5. Restart: '/?indexName[query]=Apple'
     // 6. Refine: '/?indexName[query]=Samsung'
 
     const pushState = jest.spyOn(window.history, 'pushState');
@@ -36,6 +29,7 @@ describe('routing back and forth to an InstantSearch instance', () => {
       searchClient: createSearchClient(),
       routing: {
         router: historyRouter({
+          cleanUrlOnDispose: true,
           writeDelay,
         }),
       },
@@ -43,7 +37,8 @@ describe('routing back and forth to an InstantSearch instance', () => {
 
     // 1. Initial: '/'
     {
-      addWidgetsAndStart(search);
+      search.addWidgets([connectSearchBox(() => {})({})]);
+      search.start();
 
       await wait(writeWait);
       expect(window.location.search).toEqual('');
@@ -61,37 +56,38 @@ describe('routing back and forth to an InstantSearch instance', () => {
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. Dispose: '/?indexName[query]=Apple'
+    // 3. Navigate: '/about'
     {
       search.dispose();
+      window.history.pushState({}, '', '/about');
 
       await wait(writeWait);
-      expect(window.location.search).toEqual(
-        `?${encodeURI('indexName[query]=Apple')}`
-      );
-      expect(pushState).toHaveBeenCalledTimes(1);
+      expect(window.location.pathname).toEqual('/about');
+      expect(window.location.search).toEqual('');
+      expect(pushState).toHaveBeenCalledTimes(2);
     }
 
-    // 4. Refine: '/?indexName[query]=Apple'
+    // 4. Back to previous page
     {
-      search.renderState.indexName.searchBox!.refine('Samsung');
+      window.history.back();
 
       await wait(writeWait);
+      expect(window.location.pathname).toEqual('/');
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Apple')}`
       );
-      expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 5. Start: '/?indexName[query]=Apple'
+    // 5. Restart InstantSearch
     {
-      addWidgetsAndStart(search);
+      search.addWidgets([connectSearchBox(() => {})({})]);
+      search.start();
 
       await wait(writeWait);
+      expect(window.location.pathname).toEqual('/');
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Apple')}`
       );
-      expect(pushState).toHaveBeenCalledTimes(1);
     }
 
     // 6. Refine: '/?indexName[query]=Samsung'
@@ -102,7 +98,7 @@ describe('routing back and forth to an InstantSearch instance', () => {
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Samsung')}`
       );
-      expect(pushState).toHaveBeenCalledTimes(2);
+      expect(pushState).toHaveBeenCalledTimes(3);
     }
   });
 });
