+++ packages/instantsearch-core/src/routing/__tests__/routing/correct-urls.test.ts
@@ -1,0 +1,87 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import { createSearchClient } from '@instantsearch/mocks';
+import { wait } from '@instantsearch/testutils/wait';
+
+import {
+  instantsearch,
+  index,
+  historyRouter,
+  connectPagination,
+  connectSearchBox,
+} from '../../..';
+
+beforeEach(() => {
+  window.history.pushState({}, '', '/');
+});
+
+const writeDelay = 10;
+const writeWait = 10 * writeDelay;
+
+test('correct URL for widgets', async () => {
+  const router = historyRouter({ writeDelay });
+
+  const indexName = 'indexName';
+  const search = instantsearch({
+    indexName,
+    searchClient: createSearchClient(),
+    routing: {
+      router,
+    },
+  });
+
+  search.addWidgets([
+    connectPagination(() => {})({}),
+    connectSearchBox(() => {})({}),
+  ]);
+
+  search.start();
+
+  expect(window.location.search).toBe('');
+
+  // on nested index
+  search.renderState[indexName].searchBox!.refine('test');
+  // on main index
+  search.renderState[indexName].pagination!.refine(39);
+
+  await wait(writeWait);
+
+  expect(window.location.search).toBe(
+    `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
+  );
+});
+
+test('correct URL for widgets in indices with repeated indexId', async () => {
+  const router = historyRouter({ writeDelay });
+
+  const indexName = 'indexName';
+  const search = instantsearch({
+    indexName,
+    searchClient: createSearchClient(),
+    routing: {
+      router,
+    },
+  });
+
+  search.addWidgets([
+    connectPagination(() => {})({}),
+    index({ indexName }).addWidgets([connectSearchBox(() => {})({})]),
+  ]);
+
+  search.start();
+
+  expect(window.location.search).toBe('');
+
+  // on nested index
+  search.renderState[indexName].searchBox!.refine('test');
+  // on main index
+  search.renderState[indexName].pagination!.refine(39);
+
+  await wait(writeWait);
+
+  expect(window.location.search).toBe(
+    `?${encodeURI('indexName[page]=40&indexName[query]=test')}`
+  );
+});
