+++ packages/instantsearch-core/src/routing/__tests__/routing/dispose-start.test.ts
@@ -5,9 +5,9 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import { instantsearch, historyRouter, connectSearchBox } from '../..';
+import { instantsearch, historyRouter, connectSearchBox } from '../../..';
 
-import type { InstantSearch } from '../..';
+import type { InstantSearch } from '../../..';
 
 /* eslint no-lone-blocks: "off" */
 
@@ -24,9 +24,9 @@ describe('routing back and forth to an InstantSearch instance', () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. Dispose: '/'
-    // 4. Refine: '/'
-    // 5. Start: '/'
+    // 3. Dispose: '/?indexName[query]=Apple'
+    // 4. Refine: '/?indexName[query]=Apple'
+    // 5. Start: '/?indexName[query]=Apple'
     // 6. Refine: '/?indexName[query]=Samsung'
 
     const pushState = jest.spyOn(window.history, 'pushState');
@@ -37,7 +37,6 @@ describe('routing back and forth to an InstantSearch instance', () => {
       routing: {
         router: historyRouter({
           writeDelay,
-          cleanUrlOnDispose: true,
         }),
       },
     });
@@ -62,31 +61,37 @@ describe('routing back and forth to an InstantSearch instance', () => {
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. Dispose: '/'
+    // 3. Dispose: '/?indexName[query]=Apple'
     {
       search.dispose();
 
       await wait(writeWait);
-      expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(2);
+      expect(window.location.search).toEqual(
+        `?${encodeURI('indexName[query]=Apple')}`
+      );
+      expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 4. Refine: '/'
+    // 4. Refine: '/?indexName[query]=Apple'
     {
       search.renderState.indexName.searchBox!.refine('Samsung');
 
       await wait(writeWait);
-      expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(2);
+      expect(window.location.search).toEqual(
+        `?${encodeURI('indexName[query]=Apple')}`
+      );
+      expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 5. Start: '/'
+    // 5. Start: '/?indexName[query]=Apple'
     {
       addWidgetsAndStart(search);
 
       await wait(writeWait);
-      expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(2);
+      expect(window.location.search).toEqual(
+        `?${encodeURI('indexName[query]=Apple')}`
+      );
+      expect(pushState).toHaveBeenCalledTimes(1);
     }
 
     // 6. Refine: '/?indexName[query]=Samsung'
@@ -97,7 +102,7 @@ describe('routing back and forth to an InstantSearch instance', () => {
       expect(window.location.search).toEqual(
         `?${encodeURI('indexName[query]=Samsung')}`
       );
-      expect(pushState).toHaveBeenCalledTimes(3);
+      expect(pushState).toHaveBeenCalledTimes(2);
     }
   });
 });
