+++ packages/instantsearch-core/src/routing/__tests__/routing/spa-replace-state.test.ts
@@ -5,22 +5,24 @@
 import { createSearchClient } from '@instantsearch/mocks';
 import { wait } from '@instantsearch/testutils/wait';
 
-import { instantsearch, historyRouter, connectSearchBox } from '../..';
+import { instantsearch, historyRouter, connectSearchBox } from '../../..';
 
 /* eslint no-lone-blocks: "off" */
 
 const writeDelay = 10;
 const writeWait = 10 * writeDelay;
 
-describe('routing with debounced third-party client-side router', () => {
+describe('routing using `replaceState`', () => {
+  // We can't assert whether another router did update the URL
+  // So there's no way to prevent `write` after `dispose`
   test('does not clean the URL after navigating', async () => {
     // -- Flow
     // 1. Initial: '/'
     // 2. Refine: '/?indexName[query]=Apple'
-    // 3. Dispose: '/'
-    // 4. Route change: '/about'
-    // 5. Back: '/'
-    // 6. Back: '/?indexName[query]=Apple'
+    // 3. Dispose: does not yet write
+    // 4. Route change (with `replaceState`): '/about?external=true', replaces state 2
+    // 5. Dispose: does not write
+    // 6. Back: '/'
 
     const pushState = jest.spyOn(window.history, 'pushState');
 
@@ -30,7 +32,6 @@ describe('routing with debounced third-party client-side router', () => {
       routing: {
         router: historyRouter({
           writeDelay,
-          cleanUrlOnDispose: true,
         }),
       },
     });
@@ -56,45 +57,31 @@ describe('routing with debounced third-party client-side router', () => {
       expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 3. Dispose: '/'
+    // 3. Dispose: '/about'
+    // 4. Route change (with `replaceState`): '/about'
     {
       search.dispose();
+      window.history.replaceState({}, '', '/about?external=true');
 
-      await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
-      expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(2);
-    }
+      // Asserting `replaceState` call
+      expect(window.location.pathname).toEqual('/about');
+      expect(window.location.search).toEqual('?external=true');
+      expect(pushState).toHaveBeenCalledTimes(1);
 
-    // 4. Route change: '/about'
-    {
-      window.history.pushState({}, '', '/about');
-
+      // Asserting `dispose` calling `pushState`
       await wait(writeWait);
       expect(window.location.pathname).toEqual('/about');
-      expect(window.location.search).toEqual('');
-      expect(pushState).toHaveBeenCalledTimes(3);
+      expect(window.location.search).toEqual('?external=true');
+      expect(pushState).toHaveBeenCalledTimes(1);
     }
 
-    // 5. Back: '/'
+    // 5. Back: '/about'
     {
       window.history.back();
 
       await wait(writeWait);
       expect(window.location.pathname).toEqual('/');
       expect(window.location.search).toEqual('');
-    }
-
-    // 6. Back: '/?indexName[query]=Apple'
-    {
-      window.history.back();
-
-      await wait(writeWait);
-      expect(window.location.pathname).toEqual('/');
-      expect(window.location.search).toEqual(
-        `?${encodeURI('indexName[query]=Apple')}`
-      );
-      expect(pushState).toHaveBeenCalledTimes(3);
     }
   });
 });
